Fix all security issues and upgrade to latest version with test passed
