+++ package.json
@@ -13,28 +13,32 @@
     "lint": "next lint --fix"
   },
   "dependencies": {
-    "@tanstack/react-query": "^5.17.19",
-    "@wix/dashboard": "^1.3.7",
-    "@wix/design-system": "^1.85.0",
-    "@wix/ecom": "^1.0.603",
-    "@wix/sdk": "^1.10.2",
-    "@wix/wix-ui-icons-common": "^3.16.0",
-    "jsonwebtoken": "^9.0.2",
-    "next": "14.0.4",
-    "react": "^18",
-    "react-dom": "^18"
+    "@tanstack/react-query": "^5.90.21",
+    "@wix/dashboard": "^1.3.43",
+    "@wix/design-system": "^1.257.0",
+    "@wix/ecom": "^1.0.1889",
+    "@wix/sdk": "^1.21.3",
+    "@wix/wix-ui-icons-common": "^3.139.0",
+    "jsonwebtoken": "^9.0.3",
+    "next": "^16.1.6",
+    "react": "^19.2.4",
+    "react-dom": "^19.2.4"
   },
   "devDependencies": {
-    "@playwright/test": "^1.42.1",
-    "@types/jsonwebtoken": "^9.0.5",
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "eslint": "^8",
-    "eslint-config-next": "14.0.4",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.1.3",
-    "prettier": "^3.2.3",
-    "typescript": "^5"
+    "@playwright/test": "^1.58.2",
+    "@types/jsonwebtoken": "^9.0.10",
+    "@types/node": "^25.3.0",
+    "@types/react": "^19.2.14",
+    "@types/react-dom": "^19.2.3",
+    "eslint": "^10.0.1",
+    "eslint-config-next": "^12.0.4",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-prettier": "^5.5.5",
+    "prettier": "^3.8.1",
+    "typescript": "^5.9.3"
+  },
+  "overrides": {
+    "d3-color": ">=3.1.0",
+    "minimatch": ">=10.2.1"
   }
 }
