+++ src/app/dashboard/parts/ShippingRatesPage.tsx
@@ -2,11 +2,14 @@
 import '@wix/design-system/styles.global.css';
 import { AppProviders } from '@/app/utils/client-providers';
 import { ShippingRatesPageContent } from '@/app/dashboard/parts/ShippingRatesPageContent';
+import { Suspense } from 'react';
 
-export const ShippingRatesPage = ({}: {}) => {
+export const ShippingRatesPage = ({ }: {}) => {
   return (
     <AppProviders>
-      <ShippingRatesPageContent />
+      <Suspense fallback={<div>Loading...</div>}>
+        <ShippingRatesPageContent />
+      </Suspense>
     </AppProviders>
   );
 };
